add video player window on click
